refactor(releases): split bundle document type into shared, release and variant types
