+++ packages/sanity/src/core/releases/hooks/__tests__/useCustomReleaseActions.test.tsx
@@ -6,7 +6,7 @@ import {type Source} from '../../../config/types'
 import {useSource} from '../../../studio'
 import {activeASAPRelease} from '../../__fixtures__/release.fixture'
 import {documentsInRelease} from '../../tool/detail/__tests__/__mocks__/useBundleDocuments.mock'
-import {type DocumentInRelease} from '../../tool/detail/useBundleDocuments'
+import {type DocumentInRelease} from '../../tool/detail/types'
 import {useCustomReleaseActions} from '../useCustomReleaseActions'
 
 vi.mock('../../../studio', () => ({
