+++ packages/sanity/src/core/releases/hooks/useCustomReleaseActions.ts
@@ -3,7 +3,7 @@ import {useMemo} from 'react'
 
 import {type ReleaseActionComponent} from '../../config/releases/actions'
 import {useSource} from '../../studio'
-import {type DocumentInRelease} from '../tool/detail/useBundleDocuments'
+import {type DocumentInRelease} from '../tool/detail/types'
 
 /**
  * Hook to get custom/configured release actions for a release
