+++ packages/sanity/src/core/releases/tool/components/releaseCTAButtons/ReleaseRevertButton/usePostPublishTransactions.ts
@@ -6,7 +6,7 @@ import {useClient} from '../../../../../hooks/useClient'
 import {getTransactionsLogs} from '../../../../../store/translog/getTransactionsLogs'
 import {getPublishedId} from '../../../../../util/draftUtils'
 import {RELEASES_STUDIO_CLIENT_OPTIONS} from '../../../../util/releasesClient'
-import {type DocumentInRelease} from '../../../detail/useBundleDocuments'
+import {type DocumentInRelease} from '../../../detail/types'
 
 export const usePostPublishTransactions = (documents: DocumentInRelease[]) => {
   const client = useClient(RELEASES_STUDIO_CLIENT_OPTIONS)
