+++ packages/sanity/src/core/releases/tool/detail/getPublishedArchivedReleaseDocumentsObservable.ts
@@ -6,7 +6,8 @@ import {catchError, expand, finalize, map, reduce, shareReplay} from 'rxjs/opera
 import {type useSource} from '../../../studio'
 import {getReleaseIdFromReleaseDocumentId} from '../../util/getReleaseIdFromReleaseDocumentId'
 import {RELEASES_STUDIO_CLIENT_OPTIONS} from '../../util/releasesClient'
-import {type BundleDocumentsObservableResult, type DocumentInRelease} from './useBundleDocuments'
+import {type DocumentInRelease} from './types'
+import {type BundleDocumentsObservableResult} from './useBundleDocuments'
 
 const publishedArchivedReleaseDocumentsCache: Record<string, BundleDocumentsObservableResult> =
   Object.create(null)
