+++ packages/sanity/src/core/releases/tool/detail/types.ts
@@ -1,0 +1,9 @@
+import {type BundleDocument} from './useBundleDocuments'
+
+/**
+ * A single document in a release, along with its validation status.
+ * Returned by {@link useReleaseDocuments}.
+ *
+ * @internal
+ */
+export type DocumentInRelease = BundleDocument
